chore(core-manager): extract Stream objects
